Fix create

+++ capnp/create.response.capnp
@@ -8,7 +8,7 @@ struct CreateResponse {
   attr   @4 :Attr;
 
   struct FuseFileInfo {
-    flags         @0  :Int8;
+    flags         @0  :Int64;
     writepage     @1  :UInt64;
     directIo      @2  :UInt64;
     keepCache     @3  :UInt64;
